Use size_t/ssize_t data types where appropiate

+++ src/server/reactor_thread.cc
@@ -826,8 +826,8 @@ static int ReactorThread_init(Server *serv, Reactor *reactor, uint16_t reactor_i
     reactor->max_socket = serv->get_max_connection();
     reactor->close = Server::close_connection;
 
-    reactor->set_exit_condition(Reactor::EXIT_CONDITION_DEFAULT, [thread](Reactor *reactor, int &event_num) -> bool {
-        return event_num == (int) thread->pipe_num;
+    reactor->set_exit_condition(Reactor::EXIT_CONDITION_DEFAULT, [thread](Reactor *reactor, size_t &event_num) -> bool {
+        return event_num == (size_t) thread->pipe_num;
     });
 
     reactor->default_error_handler = ReactorThread_onClose;
